Moving new function
Moved new downloading function into the old module, removed unnecassary files
